Broken project. Working on login functionality
